@pcrespov review: rename nodes.py to project_nodes.py
