+++ packages/models-library/src/models_library/project_nodes.py
