+++ services/director-v2/src/simcore_service_director_v2/models/domains/comp_tasks.py
@@ -4,7 +4,7 @@ from uuid import UUID
 
 from models_library.constants import VERSION_RE
 from models_library.projects import ProjectID, RunningState
-from models_library.nodes import NodeID, Inputs, Outputs
+from models_library.project_nodes import NodeID, Inputs, Outputs
 from models_library.services import KEY_RE, ServiceInputs, ServiceOutputs
 from pydantic import BaseModel, Extra, Field, AnyHttpUrl, constr, validator
 from pydantic.types import PositiveInt
