+++ services/web/server/tests/unit/with_dbs/medium/test_resource_manager.py
@@ -18,7 +18,7 @@ from pytest_simcore.helpers.utils_assert import assert_status
 from pytest_simcore.helpers.utils_login import LoggedUser
 from pytest_simcore.helpers.utils_projects import NewProject
 
-from models_library.nodes import RunningState
+from models_library.project_nodes import RunningState
 from servicelib.application import create_safe_application
 from simcore_service_webserver.db import setup_db
 from simcore_service_webserver.director import setup_director
